+++ db/migrate/20220323200324_create_photos.rb
@@ -2,8 +2,8 @@ class CreatePhotos < ActiveRecord::Migration[7.0]
   def change
     create_table :photos do |t|
       t.string :picture
-      t.integer :user_id
-      t.integer :micropost_id
+      t.bigint :user_id
+      t.bigint :micropost_id
 
       t.timestamps
     end
